deployment-scripts/requirements.txt
Babel==1.3
Django==1.6.5
Fabric==1.9.0
South==1.0
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
boto==2.31.1
celery==3.1.13
ecdsa==0.11
gunicorn==19.1.0
iso8601==0.1.10
kombu==3.0.21
paramiko==1.14.0
pbr==0.10.0
prettytable==0.7.2
pycrypto==2.6.1
python-novaclient==2.18.1
pytz==2014.4
requests==2.3.0
selenium==2.42.1
simplejson==3.6.2
six==1.7.3
wsgiref==0.1.2
